requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
pillow==11.1.0
pycparser==2.22
pycryptodome==3.21.0
sqlparse==0.5.3
tzdata==2025.1
